requirements.txt
anyio==3.5.0
appdirs==1.4.4
appnope==0.1.2
asgiref==3.5.0
astmonkey==0.3.6
astor==0.8.1
attrs==21.4.0
Automat==20.2.0
autopep8==1.5.6
backcall==0.2.0
backports.entry-points-selectable==1.1.1
bcrypt==3.2.0
beautifulsoup4==4.10.0
beautifultable==1.0.1
bleach==4.1.0
bs4==0.0.1
bytecode==0.13.0
CacheControl==0.12.10
cachetools==4.2.4
cachy==0.3.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.10
cleo==0.8.1
click==8.0.3
clikit==0.6.2
colorama==0.4.4
commonmark==0.9.1
constantly==15.1.0
cors==1.0.1
crashtest==0.3.1
cryptography==36.0.1
database-connection==0.1.1
decorator==5.0.7
defusedxml==0.7.1
distlib==0.3.3
entrypoints==0.4
env==0.1.0
fastapi==0.72.0
filelock==3.4.0
firebase-admin==5.2.0
Flask-API==2.0
Flask-Cors==3.0.10
future==0.18.2
GDAL==3.4.1
gevent==21.12.0
google-api-core==1.31.4
google-api-python-client==2.32.0
google-auth==1.35.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==1.7.2
google-cloud-firestore==2.3.4
google-cloud-language==2.3.1
google-cloud-storage==2.1.0
google-cloud-translate==2.0.1
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.53.0
googletrans==3.1.0a0
greenlet==1.1.2
grpcio==1.42.0
gunicorn==20.1.0
h11==0.13.0
h2==3.2.0
hpack==3.0.0
hstspreload==2021.12.1
html5lib==1.1
httpcore==0.9.1
httplib2==0.20.2
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.11.0
incremental==21.3.0
iniconfig==1.1.1
ipykernel==5.5.5
ipython==7.23.1
ipython-genutils==0.2.0
jedi==0.18.0
jellyfish==0.9.0
Jinja2==3.0.3
jsonschema==4.4.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
keyring==23.5.0
libretranslatepy==2.1.1
lockfile==0.12.2
lxml==4.6.4
MarkupSafe==2.0.1
matplotlib-inline==0.1.2
mistune==0.8.4
msgpack==1.0.2
MutPy-Pynguin==0.7.1
mypy-extensions==0.4.3
mysql-connector-python==8.0.27
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
neovim==0.3.1
nest-asyncio==1.5.4
networkx==2.6.3
notebook-as-pdf==0.5.0
numpy==1.22.3
ordered-set==4.1.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.5.0
parso==0.8.2
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pickleshare==0.7.5
pip-chill==1.0.1
pkginfo==1.8.2
platformdirs==2.4.0
pluggy==1.0.0
poetry==1.1.13
poetry-core==1.0.7
powerline-status==2.7
prompt-toolkit==3.0.18
proto-plus==1.19.8
protobuf==3.19.1
psycopg2==2.9.3
ptyprocess==0.7.0
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.21
pydantic==1.9.0
pydot==1.4.2
pyee==8.2.2
Pygments==2.11.2
PyJWT==2.3.0
pylev==1.4.0
PyMySQL==1.0.2
pynguin==0.17.0
pynvim==0.4.3
pyparsing==3.0.7
PyPDF2==1.26.0
pyppeteer==1.0.2
pyrsistent==0.18.1
PySocks==1.7.1
pytest==6.2.5
python-dateutil==2.8.2
python-dotenv==0.19.2
python-multipart==0.0.5
pytz==2021.1
PyYAML==6.0
pyzmq==22.0.3
requests==2.27.1
requests-file==1.5.1
requests-toolbelt==0.9.1
rfc3986==1.5.0
rich==11.2.0
rsa==4.8
shellingham==1.4.0
simple-parsing==0.0.17
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
sourcery-cli==0.10.1
sql==0.4.0
sql-alchemy-db==0.0.4
SQLAlchemy==1.4.31
sqlalchemy-redshift==0.8.9
starlette==0.17.1
termcolor==1.1.0
testpath==0.5.0
tldextract==3.2.0
toml==0.10.2
tomlkit==0.9.2
tornado==6.1
tqdm==4.62.3
traitlets==5.0.5
translate==3.6.1
Twisted==21.7.0
typing-inspect==0.7.1
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
uvicorn==0.17.0
virtualenv==20.10.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==10.1
zipp==3.7.0
zope.event==4.5.0
zope.interface==5.4.0
